+++ common/constants/dev.ts
@@ -1,0 +1,6 @@
+// 开发者账号
+export const DEVELOPER = {
+  ID: '5ba9a4a4b1493d4fcc920efc',
+  NAME: 'zhuxiankang'
+}
+