+++ graphql/lokka.ts
@@ -1,7 +1,8 @@
 import { Transport } from 'lokka-transport-http'
 import { Lokka } from 'lokka'
-import config from '~/nuxt.config'
+
+console.log(`🚀🚀🚀 Client is requesting on ${process.env.PROXY_HTTP}`)
 
 export const client = new Lokka({
-  transport: new Transport(config.proxyHttp)
+  transport: new Transport(process.env.PROXY_HTTP)
 })