+++ plugins/museui.js
@@ -2,8 +2,10 @@ import Vue from 'vue'
 import MuseUI from 'muse-ui'
 import 'muse-ui/dist/muse-ui.css'
 import 'muse-ui-loading/dist/muse-ui-loading.css'
+import 'muse-ui-message/dist/muse-ui-message.css'
 import Toast from 'muse-ui-toast'
 import Loading from 'muse-ui-loading'
+import Message from 'muse-ui-message'
 
 Vue.use(MuseUI)
 
@@ -20,5 +22,7 @@ const toastConfig = {
 
 Vue.use(Toast, toastConfig)
 Vue.use(Loading)
+
+Vue.use(Message)
 
 export default toastConfig