+++ CHANGELOG.md
@@ -1,4 +1,61 @@
 <a name="1.0.0"></a>
+# 1.0.0 (2018-10-19)
+
+
+### Bug Fixes
+
+* **muse-ui.common:** tooltip报错处理，并给muse ui提了issue ([e085778](https://github.com/EmptyCup20/easy-tool/commit/e085778))
+* **test:** 测试 ([86fa1ad](https://github.com/EmptyCup20/easy-tool/commit/86fa1ad))
+* **test:** 环境测试 ([53ff7e7](https://github.com/EmptyCup20/easy-tool/commit/53ff7e7))
+* **多语言模块[Node]:** 修复语法 ([9c7f082](https://github.com/EmptyCup20/easy-tool/commit/9c7f082))
+* **多语言模块[Vue]:** 修复缺陷 ([44946a2](https://github.com/EmptyCup20/easy-tool/commit/44946a2))
+* **用户模块[Node]:** 修复用户登录缺陷 ([084c8b6](https://github.com/EmptyCup20/easy-tool/commit/084c8b6))
+* **缺陷修复:** 修复JSON.parse没有使用try catch问题 ([7ea3fcc](https://github.com/EmptyCup20/easy-tool/commit/7ea3fcc))
+* **静态资源模块[Node]:** 修复多余逻辑问题 ([b0951f9](https://github.com/EmptyCup20/easy-tool/commit/b0951f9))
+* **静态资源模块[Vue]:** 修改muse-ui的版本为3.0.0 ([1b833f7](https://github.com/EmptyCup20/easy-tool/commit/1b833f7))
+* **项目模块[Vue]:** 对muse-ui的多选组件填充数据进行注释说明 ([77cb250](https://github.com/EmptyCup20/easy-tool/commit/77cb250))
+
+
+### Features
+
+* **font-awesome:** 新增图标库 ([0a0234f](https://github.com/EmptyCup20/easy-tool/commit/0a0234f))
+* **nuxt.config.js:** 修改PORT和HOST全局配置 ([7dc39bd](https://github.com/EmptyCup20/easy-tool/commit/7dc39bd))
+* **package.json:** 修改pro和dev的启动命令 ([086fa69](https://github.com/EmptyCup20/easy-tool/commit/086fa69))
+* **pages/index:** 新增登录页面 ([32d0818](https://github.com/EmptyCup20/easy-tool/commit/32d0818))
+* **plugins:** 支持muse-ui插件 ([92037db](https://github.com/EmptyCup20/easy-tool/commit/92037db))
+* **server:** 去掉服务中间件api ([9e670f9](https://github.com/EmptyCup20/easy-tool/commit/9e670f9))
+* **server:** 新增ts-node ([af8dba6](https://github.com/EmptyCup20/easy-tool/commit/af8dba6))
+* **server:** 新增服务端跨域，用于解决前后端分离开发的问题，nuxt.config.js新增proxy_http代理服务地址 ([3f296b8](https://github.com/EmptyCup20/easy-tool/commit/3f296b8))
+* **server:** 服务端ts-node调试完成 ([bd99a55](https://github.com/EmptyCup20/easy-tool/commit/bd99a55))
+* **server/plugins/package.json:** 前端新增代理配置，后端新增跨域，实现前后端开发分离 ([631812b](https://github.com/EmptyCup20/easy-tool/commit/631812b))
+* **多语言模块[Node]:** excel多语言导入完成 ([b44953d](https://github.com/EmptyCup20/easy-tool/commit/b44953d))
+* **多语言模块[Node]:** excel的多语言导入完成 ([6261419](https://github.com/EmptyCup20/easy-tool/commit/6261419))
+* **多语言模块[Node]:** 上传多语言 ([5742628](https://github.com/EmptyCup20/easy-tool/commit/5742628))
+* **多语言模块[Node]:** 多语言增加和查询功能完成 ([0913641](https://github.com/EmptyCup20/easy-tool/commit/0913641))
+* **多语言模块[Node]:** 多语言导入导出完成 ([de34815](https://github.com/EmptyCup20/easy-tool/commit/de34815))
+* **多语言模块[Node]:** 多语言导入建设中 ([b62de10](https://github.com/EmptyCup20/easy-tool/commit/b62de10))
+* **多语言模块[Node]:** 多语言导入支持json、xlsx以及properties导入功能完成 ([3d04cab](https://github.com/EmptyCup20/easy-tool/commit/3d04cab))
+* **多语言模块[Node]:** 多语言导入设计中 ([674040f](https://github.com/EmptyCup20/easy-tool/commit/674040f))
+* **多语言模块[Node]:** 导入多语言列表完成 ([5f7bbe5](https://github.com/EmptyCup20/easy-tool/commit/5f7bbe5))
+* **多语言模块[Node]:** 翻译多语言上传处理 ([581d377](https://github.com/EmptyCup20/easy-tool/commit/581d377))
+* **多语言模块[Node]:** 翻译导入完成 ([4b38a53](https://github.com/EmptyCup20/easy-tool/commit/4b38a53))
+* **用户模块[Node]:** 新增用户session ([74194e2](https://github.com/EmptyCup20/easy-tool/commit/74194e2))
+* **用户模块[Node]:** 生产态sesssion设置 ([0398542](https://github.com/EmptyCup20/easy-tool/commit/0398542))
+* **用户模块[Node]:** 用户登录 ([53d4f8f](https://github.com/EmptyCup20/easy-tool/commit/53d4f8f))
+* **用户模块[Node]:** 用户登录请求设计中 ([6fd9f27](https://github.com/EmptyCup20/easy-tool/commit/6fd9f27))
+* **配置项目:** 从Bitbucket切回Github ([3954631](https://github.com/EmptyCup20/easy-tool/commit/3954631))
+* **静态资源模块[Node]:** 删除项目资源类型时删除资源数据 ([5cedd1f](https://github.com/EmptyCup20/easy-tool/commit/5cedd1f))
+* **静态资源模块[Node]:** 新增静态资源类型 ([5408e81](https://github.com/EmptyCup20/easy-tool/commit/5408e81))
+* **静态资源模块[Node]:** 资源模块增删改查成功 ([69a523c](https://github.com/EmptyCup20/easy-tool/commit/69a523c))
+* **静态资源模块[Node]:** 静态资源数据模型构建 ([77759b5](https://github.com/EmptyCup20/easy-tool/commit/77759b5))
+* **项目模块[Node]:** 查询项目列表和用户列表完成 ([5b923b4](https://github.com/EmptyCup20/easy-tool/commit/5b923b4))
+* **项目模块[Node]:** 用户登录 ([8ced17f](https://github.com/EmptyCup20/easy-tool/commit/8ced17f))
+* **项目模块[Node]:** 项目增删改查完成 ([5046699](https://github.com/EmptyCup20/easy-tool/commit/5046699))
+* **项目模块[Vue]:** 项目的增加和修改建设中 ([ff1688d](https://github.com/EmptyCup20/easy-tool/commit/ff1688d))
+
+
+
+<a name="1.0.0"></a>
 # 1.0.0 (2018-08-30)